Revert Nitro type change.
And correct wrong variable name for MvsLMenu.

+++ include/nitro_if.h
@@ -50,13 +50,13 @@ typedef void* NITRO_TYPE_REPLACEMENT;
 	void MI_CpuCopy8(const void*, void*, int);
 
 	typedef unsigned char u8;
-	typedef unsigned short u16;
-	typedef unsigned int u32;
+	typedef unsigned short int u16;
+	typedef unsigned long u32;
 	typedef unsigned long long int u64;
 
 	typedef signed char s8;
-	typedef signed short s16;
-	typedef signed int s32;
+	typedef signed short int s16;
+	typedef signed long s32;
 	typedef signed long long int s64;
 
 	typedef s16 fx16;
